+++ tests/main.cpp
@@ -4,6 +4,7 @@
 
 #include "list_test.hpp"
 #include "spin_lock_test.hpp"
+#include "mpsc_queue_test.hpp"
 #include "shared_test.hpp"
 #include "event_test.hpp"
 #include "thread_test.hpp"